edit-integer-to-long

+++ src/main/java/com/example/library/Repository/BookRepo.java
@@ -3,5 +3,5 @@ package com.example.library.Repository;
 import com.example.library.model.entity.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookRepo extends JpaRepository<Book, Integer> {
+public interface BookRepo extends JpaRepository<Book, Long> {
 }
