+++ src/main/java/com/example/library/Repository/ReservRepo.java
@@ -3,5 +3,5 @@ package com.example.library.Repository;
 import com.example.library.model.entity.Reserv;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReservRepo extends JpaRepository<Reserv, Integer> {
+public interface ReservRepo extends JpaRepository<Reserv, Long> {
 }
